requirements.txt
cryptography
Flask
argon2-cffi